requirements.txt
# Core dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# LLM dependencies - Ollama for self-hosted LLaMA
ollama>=0.1.0

# Vector database
pinecone-client>=3.0.0

# Embeddings
sentence-transformers>=2.3.1
huggingface-hub>=0.34.0,<1.0

# UI
streamlit>=1.31.0

# Utilities
httpx>=0.26.0
python-multipart>=0.0.6
aiofiles>=23.2.1
tiktoken>=0.5.2
requests>=2.31.0
